Add SendFundsHandlingApi, remove HDWalletApi
