+++ wallet-test/src/test/scala/org/bitcoins/wallet/FundTransactionHandlingTest.scala
@@ -1,10 +1,10 @@
 package org.bitcoins.wallet
 
-import org.bitcoins.core.api.wallet.HDWalletApi
+import org.bitcoins.core.api.wallet.NeutrinoHDWalletApi
 import org.bitcoins.core.currency.Bitcoins
 import org.bitcoins.core.protocol.transaction.TransactionOutput
 import org.bitcoins.core.wallet.utxo.StorageLocationTag.HotStorage
-import org.bitcoins.core.wallet.utxo._
+import org.bitcoins.core.wallet.utxo.*
 import org.bitcoins.testkit.wallet.{
   BitcoinSWalletTestCachedBitcoindNewest,
   WalletTestUtil,
@@ -218,8 +218,8 @@ class FundTransactionHandlingTest
       val bitcoind = fundedWallet.bitcoind
       val fundedTxF = for {
         feeRate <- wallet.getFeeRate()
-        _ <- wallet.accountHandling.createNewAccount(
-          wallet.keyManager.kmParams.purpose)
+//        _ <- wallet.accountHandling.createNewAccount(
+//          wallet.kmParams.purpose)
         accounts <- wallet.accountHandling.listAccounts()
         account2 = accounts.find(_.hdAccount.index == 2).get
 
@@ -269,14 +269,16 @@ class FundTransactionHandlingTest
   }
 
   def testAddressTagFunding(
-      wallet: HDWalletApi,
+      wallet: NeutrinoHDWalletApi,
       tag: AddressTag
   ): Future[Assertion] = {
     for {
       feeRate <- wallet.getFeeRate()
       taggedAddr <- wallet.addressHandling.getNewAddress(Vector(tag))
       _ <-
-        wallet.sendToAddress(taggedAddr, destination.value * 2, Some(feeRate))
+        wallet.sendFundsHandling.sendToAddress(taggedAddr,
+                                               destination.value * 2,
+                                               Some(feeRate))
       taggedBalance <- wallet.getBalance(tag)
       _ = assert(taggedBalance == destination.value * 2)
