+++ wallet/src/main/scala/org/bitcoins/wallet/WalletHolder.scala
@@ -9,7 +9,6 @@ import org.bitcoins.core.api.dlc.wallet.db.{
   IncomingDLCOfferDb
 }
 import org.bitcoins.core.api.feeprovider.FeeRateApi
-import org.bitcoins.core.api.keymanager.BIP39KeyManagerApi
 import org.bitcoins.core.api.node.NodeApi
 import org.bitcoins.core.api.wallet.*
 import org.bitcoins.core.api.wallet.db.*
@@ -27,7 +26,6 @@ import org.bitcoins.core.protocol.transaction.{
   TransactionOutput
 }
 import org.bitcoins.core.protocol.BitcoinAddress
-import org.bitcoins.core.psbt.PSBT
 import org.bitcoins.core.wallet.fee.{FeeUnit, SatoshisPerVirtualByte}
 import org.bitcoins.core.wallet.utxo.{AddressTag, TxoState}
 import org.bitcoins.crypto.{DoubleSha256DigestBE, Sha256Digest}
@@ -68,6 +66,9 @@ class WalletHolder(initWalletOpt: Option[DLCNeutrinoHDWalletApi])(implicit
 
   override def transactionProcessing: TransactionProcessingApi =
     wallet.transactionProcessing
+
+  override def sendFundsHandling: SendFundsHandlingApi =
+    wallet.sendFundsHandling
   def isInitialized: Boolean = synchronized {
     walletOpt.isDefined
   }
@@ -184,27 +185,6 @@ class WalletHolder(initWalletOpt: Option[DLCNeutrinoHDWalletApi])(implicit
   ): Future[Vector[SpendingInfoDb]] = delegate(_.unmarkUTXOsAsReserved(tx))
 
   override def isEmpty(): Future[Boolean] = delegate(_.isEmpty())
-
-  override def sendFromOutPoints(
-      outPoints: Vector[TransactionOutPoint],
-      address: BitcoinAddress,
-      feeRate: FeeUnit
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendFromOutPoints(outPoints, address, feeRate))
-
-  override def sweepWallet(address: BitcoinAddress, feeRate: FeeUnit)(implicit
-      ec: ExecutionContext
-  ): Future[Transaction] = delegate(_.sweepWallet(address, feeRate))
-
-  override def bumpFeeRBF(
-      txId: DoubleSha256DigestBE,
-      newFeeRate: FeeUnit
-  ): Future[Transaction] = delegate(_.bumpFeeRBF(txId, newFeeRate))
-
-  override def bumpFeeCPFP(
-      txId: DoubleSha256DigestBE,
-      feeRate: FeeUnit
-  ): Future[Transaction] = delegate(_.bumpFeeCPFP(txId, feeRate))
 
   override def isChange(output: TransactionOutput): Future[Boolean] = delegate(
     _.isChange(output)
@@ -379,75 +359,11 @@ class WalletHolder(initWalletOpt: Option[DLCNeutrinoHDWalletApi])(implicit
       address: InetSocketAddress
   ): Future[Vector[DLCStatus]] = delegate(_.listDLCsByContact(address))
 
-  override def keyManager: BIP39KeyManagerApi = wallet.keyManager
-
   def getConfirmedBalance(account: HDAccount): Future[CurrencyUnit] =
     delegate(_.accountHandling.getConfirmedBalance(account))
 
   def getUnconfirmedBalance(account: HDAccount): Future[CurrencyUnit] =
     delegate(_.accountHandling.getUnconfirmedBalance(account))
-
-  override def sendWithAlgo(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      algo: CoinSelectionAlgo,
-      fromAccount: AccountDb,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendWithAlgo(address, amount, feeRate, algo, fromAccount, newTags)
-  )
-
-  override def sendFromOutPoints(
-      outPoints: Vector[TransactionOutPoint],
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      fromAccount: AccountDb,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendFromOutPoints(
-      outPoints,
-      address,
-      amount,
-      feeRate,
-      fromAccount,
-      newTags
-    )
-  )
-
-  override def sendToAddress(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      fromAccount: AccountDb,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendToAddress(address, amount, feeRate, fromAccount, newTags)
-  )
-
-  override def sendToAddresses(
-      addresses: Vector[BitcoinAddress],
-      amounts: Vector[CurrencyUnit],
-      feeRate: FeeUnit,
-      fromAccount: AccountDb,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendToAddresses(addresses, amounts, feeRate, fromAccount, newTags)
-  )
-
-  override def sendToOutputs(
-      outputs: Vector[TransactionOutput],
-      feeRate: FeeUnit,
-      fromAccount: AccountDb,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendToOutputs(outputs, feeRate, fromAccount, newTags)
-  )
-
-  override def signPSBT(psbt: PSBT)(implicit
-      ec: ExecutionContext
-  ): Future[PSBT] = delegate(_.signPSBT(psbt))
 
   override def getSyncDescriptorOpt(): Future[Option[SyncHeightDescriptor]] =
     delegate(_.getSyncDescriptorOpt())
@@ -473,66 +389,6 @@ class WalletHolder(initWalletOpt: Option[DLCNeutrinoHDWalletApi])(implicit
       ec: ExecutionContext
   ): Future[CurrencyUnit] = delegate(_.getBalance(tag))
 
-  override def sendFromOutPoints(
-      outPoints: Vector[TransactionOutPoint],
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRateOpt: Option[FeeUnit]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendFromOutPoints(outPoints, address, amount, feeRateOpt)
-  )
-
-  override def sendFromOutPoints(
-      outPoints: Vector[TransactionOutPoint],
-      address: BitcoinAddress,
-      feeRateOpt: Option[FeeUnit]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendFromOutPoints(outPoints, address, feeRateOpt)
-  )
-
-  override def sweepWallet(address: BitcoinAddress)(implicit
-      ec: ExecutionContext
-  ): Future[Transaction] = delegate(_.sweepWallet(address))
-
-  override def sweepWallet(
-      address: BitcoinAddress,
-      feeRateOpt: Option[FeeUnit]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sweepWallet(address, feeRateOpt)
-  )
-
-  override def sendWithAlgo(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRateOpt: Option[FeeUnit],
-      algo: CoinSelectionAlgo
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendWithAlgo(address, amount, feeRateOpt, algo)
-  )
-
-  override def sendToAddress(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRateOpt: Option[FeeUnit]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendToAddress(address, amount, feeRateOpt)
-  )
-
-  override def sendToOutputs(
-      outputs: Vector[TransactionOutput],
-      feeRateOpt: Option[FeeUnit]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendToOutputs(outputs, feeRateOpt)
-  )
-
-  override def sendToAddresses(
-      addresses: Vector[BitcoinAddress],
-      amounts: Vector[CurrencyUnit],
-      feeRateOpt: Option[FeeUnit]
-  )(implicit ec: ExecutionContext): Future[Transaction] = delegate(
-    _.sendToAddresses(addresses, amounts, feeRateOpt)
-  )
-
   def getBalance(account: HDAccount)(implicit
       ec: ExecutionContext
   ): Future[CurrencyUnit] = delegate(_.accountHandling.getBalance(account))
@@ -542,169 +398,6 @@ class WalletHolder(initWalletOpt: Option[DLCNeutrinoHDWalletApi])(implicit
       blockFilter: GolombFilter
   ): Future[NeutrinoHDWalletApi] =
     delegate(_.processCompactFilter(blockHash, blockFilter))
-
-  override def sendWithAlgo(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      algo: CoinSelectionAlgo,
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendWithAlgo(address, amount, feeRate, algo, fromAccount))
-
-  override def sendWithAlgo(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRateOpt: Option[FeeUnit],
-      algo: CoinSelectionAlgo,
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendWithAlgo(address, amount, feeRateOpt, algo, fromAccount))
-
-  override def sendWithAlgo(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      algo: CoinSelectionAlgo
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendWithAlgo(address, amount, feeRate, algo))
-
-  override def sendWithAlgo(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      algo: CoinSelectionAlgo,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendWithAlgo(address, amount, feeRate, algo, newTags))
-
-  override def sendFromOutPoints(
-      outPoints: Vector[TransactionOutPoint],
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(
-      _.sendFromOutPoints(outPoints, address, amount, feeRate, fromAccount)
-    )
-
-  override def sendFromOutPoints(
-      outPoints: Vector[TransactionOutPoint],
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRateOpt: Option[FeeUnit],
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(
-      _.sendFromOutPoints(outPoints, address, amount, feeRateOpt, fromAccount)
-    )
-
-  override def sendFromOutPoints(
-      outPoints: Vector[TransactionOutPoint],
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendFromOutPoints(outPoints, address, amount, feeRate))
-
-  override def sendFromOutPoints(
-      outPoints: Vector[TransactionOutPoint],
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendFromOutPoints(outPoints, address, amount, feeRate, newTags))
-
-  override def sendToAddress(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToAddress(address, amount, feeRate, fromAccount))
-
-  override def sendToAddress(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRateOpt: Option[FeeUnit],
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToAddress(address, amount, feeRateOpt, fromAccount))
-
-  override def sendToAddress(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToAddress(address, amount, feeRate))
-
-  override def sendToAddress(
-      address: BitcoinAddress,
-      amount: CurrencyUnit,
-      feeRate: FeeUnit,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToAddress(address, amount, feeRate, newTags))
-
-  override def sendToAddresses(
-      addresses: Vector[BitcoinAddress],
-      amounts: Vector[CurrencyUnit],
-      feeRate: FeeUnit,
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToAddresses(addresses, amounts, feeRate, fromAccount))
-
-  override def sendToAddresses(
-      addresses: Vector[BitcoinAddress],
-      amounts: Vector[CurrencyUnit],
-      feeRateOpt: Option[FeeUnit],
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToAddresses(addresses, amounts, feeRateOpt, fromAccount))
-
-  override def sendToAddresses(
-      addresses: Vector[BitcoinAddress],
-      amounts: Vector[CurrencyUnit],
-      feeRate: FeeUnit
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToAddresses(addresses, amounts, feeRate))
-
-  override def sendToAddresses(
-      addresses: Vector[BitcoinAddress],
-      amounts: Vector[CurrencyUnit],
-      feeRate: FeeUnit,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToAddresses(addresses, amounts, feeRate, newTags))
-
-  override def sendToOutputs(
-      outputs: Vector[TransactionOutput],
-      feeRate: FeeUnit,
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToOutputs(outputs, feeRate, fromAccount))
-
-  override def sendToOutputs(
-      outputs: Vector[TransactionOutput],
-      feeRateOpt: Option[FeeUnit],
-      fromAccount: AccountDb
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToOutputs(outputs, feeRateOpt, fromAccount))
-
-  override def sendToOutputs(
-      outputs: Vector[TransactionOutput],
-      feeRate: FeeUnit,
-      newTags: Vector[AddressTag]
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToOutputs(outputs, feeRate, newTags))
-
-  override def sendToOutputs(
-      outputs: Vector[TransactionOutput],
-      feeRate: FeeUnit
-  )(implicit ec: ExecutionContext): Future[Transaction] =
-    delegate(_.sendToOutputs(outputs, feeRate))
 
   override def createDLCOffer(
       contractInfoTLV: ContractInfoTLV,
